requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
simple-salesforce==1.12.5
anthropic==0.26.0
weasyprint==61.2
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.31.0
toml==0.10.2
fpdf2==2.8.3
graphviz==0.20.3
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
pytest-mock==3.14.0
